[SPARK-14914] Fix Resource not closed after using.
This change fixes a set of issues in both test code and dev code, which
does not close resources after use. It does not cause problem on UNIX
like system, but always cause failures on Windows, because Windows does
not allow deleting opened files.

NOTE: This is a squashed change from earlier pull request from
https://github.com/apache/spark/pull/12693, rebased and squashed to
catch up the latest changes in master branch.